Remove `CommitmentDatabases` data structure into its own module with minimal changes
